Changes for V.4 and ANSI C
git-svn-id: https://svn.so-much-stuff.com/svn/trunk/cvs/trunk@1447 93276802-5317-4dee-a4ef-7ae40ebda4b9

+++ games.d/hack1.0.3.d/hack.invent.c
@@ -411,7 +411,7 @@ register char *let,*word;
 			pline("You don't have that object.");
 			continue;
 		}
-		if(cnt < 0 || otmp->quan < cnt) {
+		if(cnt < 0 || otmp->quan < (unsigned)cnt) {
 			pline("You don't have that many! [You have %u]"
 			, otmp->quan);
 			continue;
@@ -790,7 +790,7 @@ extern char *occtxt;
 static long goldcounted;
 
 countgold(){
-	if((goldcounted += 100*(u.ulevel + 1)) >= u.ugold) {
+	if((unsigned)(goldcounted += 100*(u.ulevel + 1)) >= u.ugold) {
 		long eps = 0;
 		if(!rn2(2)) eps = rnd((int) (u.ugold/100 + 1));
 		pline("You probably have about %ld gold pieces.",
